ninja windows: Support x64 configuration platform
This supports top-level x64 configuration so the entire
project can be built as 64 bit. Previously, we only supported
a 64 bit override for individual targets (via msvs_target_platform).

'asm' files were a bit confused before; the intention was that all .asm
be built on Win32, but they were only being built if the msvs_target_platform
was explicitly set which was very confusing. So, support .asm still, but if
the target overrides asm with a rule, then use that one instead, similar to
.idl files. 

BUG=chromium:8606
Review URL: https://codereview.chromium.org/11363143