Adds the ability to skip includes.
Checks for a flag set in the toplevel dictionary of a file, and if its set, the includes for this file will be skipped. In the Chrome build, this allows us to write final .gyp files with all flags preprocessed and to not modify anything based on common.gypi.

This reverts r1723 (write link deps file) since this new GN integration replaces it.

BUG=307571
Review=https://codereview.chromium.org/27353003/