android: Allow disabling the generation of alias targets.
The android backend mangles the names of targets in its output in order
to ensure they don't clash with each other or with existing Android
build system targets, since they are required to be unique. It also
generates aliases referring to their original names as given in the gyp
input, for usability and test purposes, but this can cause namespace
issues. Add a generator option allowing the generation of these aliases
to be disabled, which will also disable generation of the
"gyp_all_modules".

BUG=gyp:341
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/222243004